refactor(frontend): moved Receipt-related components into a dedicated folder
